Keywords search

+++ Library-Management-System/src/edu/cmpe275/team13/controllers/BookController.java
@@ -4,6 +4,8 @@ import java.sql.Date;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -14,10 +16,8 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import edu.cmpe275.team13.beans.Book;
-import edu.cmpe275.team13.persistence.BookDAOImpl;
-import edu.cmpe275.team13.search.BookRepository;
+import edu.cmpe275.team13.exceptions.UnauthorizedAccessException;
 import edu.cmpe275.team13.search.BookSearch;
-import edu.cmpe275.team13.search.BookSpecification;
 import edu.cmpe275.team13.service.BookService;
 
 @Controller
@@ -26,9 +26,6 @@ public class BookController {
 
 	@Autowired
 	private BookService bookservice;
-
-	@Autowired
-	private BookRepository<Book> bookRepository;
 
 	/**
 	 * @return the bookservice
@@ -53,20 +50,29 @@ public class BookController {
 	}
 
 	@RequestMapping(value = "librarian/{isbn}", method = RequestMethod.GET)
-	public String getBookLibrarian(@PathVariable long isbn, Model model) {
+	public String getBookLibrarian(@PathVariable long isbn, Model model, HttpSession session) {
+		if(!session.getAttribute("type").equals("librarian")) {
+			throw new UnauthorizedAccessException();
+		}
 		Book book = bookservice.getBookById(isbn);
 		model.addAttribute("book", book);
 		return "book/updatebook";
 	}
 
 	@RequestMapping(value = "/{isbn}", method = RequestMethod.DELETE)
-	public String deleteBook(@PathVariable Long isbn) {
+	public String deleteBook(@PathVariable Long isbn, HttpSession session) {
+		if(!session.getAttribute("type").equals("librarian")) {
+			throw new UnauthorizedAccessException();
+		}
 		this.bookservice.removeBook(isbn);
 		return "redirect:/index.jsp";
 	}
 
 	@RequestMapping(value = "/{isbn}", method = RequestMethod.POST)
-	public String updateBook(@PathVariable Long isbn, @ModelAttribute Book book) {
+	public String updateBook(@PathVariable Long isbn, @ModelAttribute Book book, HttpSession session) {
+		if(!session.getAttribute("type").equals("librarian")) {
+			throw new UnauthorizedAccessException();
+		}
 		Book book_db = this.bookservice.getBookById(isbn);
 		if (null == book_db) {
 			// return createBook(book); TODO
@@ -84,6 +90,7 @@ public class BookController {
 		book_db.setLocation_in_library(book.getLocation_in_library());
 		book_db.setNumber_of_copies(book.getNumber_of_copies());
 		book_db.setImage(book.getImage());
+		book_db.setKeywords(book.getKeywords());
 		// book_db.setUpdated_by(book.getUpdated_by()); TODO take it from
 		// currently logged in user.
 		this.bookservice.updateBook(book_db);
@@ -91,7 +98,10 @@ public class BookController {
 	}
 
 	@RequestMapping(method = RequestMethod.GET)
-	public String listBooks() {
+	public String listBooks(HttpSession session) {
+		if(!session.getAttribute("type").equals("librarian")) {
+			throw new UnauthorizedAccessException();
+		}
 		return "book/createbook";
 	}
 
@@ -101,9 +111,10 @@ public class BookController {
 		int updated_by = (map.get("updated_by") == null || map.get("updated_by").trim().length() == 0) ? Integer.MIN_VALUE : Integer.parseInt(map.get("updated_by"));
 		Long isbn = (map.get("isbn") == null || map.get("isbn").trim().length() == 0) ? null : Long.parseLong(map.get("isbn"));
 		Date year = (map.get("year_of_publication") == null || map.get("year_of_publication").trim().length() == 0) ? null : Date.valueOf(map.get("year_of_publication"));
-		boolean status = (map.get("book_status") == null || map.get("book_status").trim().length() == 0) ? false : ((map.get("book_status").trim().equalsIgnoreCase("false") ? false : true));
+		boolean status = (map.get("book_status") == null || map.get("book_status").trim().length() == 0) ? true : ((map.get("book_status").trim().equalsIgnoreCase("false") ? false : true));
+		String[] keywords = (map.get("keywords") == null || map.get("keywords").trim().length() == 0) ? null : (";" + map.get("keywords")).split(";");
 		BookSearch booksearch = new BookSearch(map.get("title").trim(), map.get("author_name").trim(), map.get("publisher_name").trim(),
-				isbn, year, status, created_by, updated_by);
+				isbn, year, status, created_by, updated_by, keywords);
 		List<Book> books = this.bookservice.searchBySpec(booksearch);
 		System.out.println(books.size());
 		model.addAttribute("books", books);
@@ -119,7 +130,10 @@ public class BookController {
 	 * creates a new book in the system
 	 */
 	@RequestMapping(method = RequestMethod.POST)
-	public String createBook(@ModelAttribute Book book) {
+	public String createBook(@ModelAttribute Book book, HttpSession session) {
+		if(!session.getAttribute("type").equals("librarian")) {
+			throw new UnauthorizedAccessException();
+		}
 		Long isbn = this.bookservice.addBook(book);
 		return "redirect:/books/" + isbn;
 	}
